+++ norddrop/ffi/bindings/norddrop.h
@@ -266,7 +266,6 @@ enum norddrop_result norddrop_reject_file(const struct norddrop *dev,
  * * `dev` - Pointer to the instance
  * * `listen_addr` - Address to listen on
  * * `config` - JSON configuration
- * * `tracker_context` - App trackers context
  *
  * # Configuration Parameters
  *
@@ -290,6 +289,8 @@ enum norddrop_result norddrop_reject_file(const struct norddrop *dev,
  * * `moose_event_path` - moose database path. It MUST NOT be the same as
  * the path used for the app tracker.
  *
+ * * `tracker_context` - App trackers context
+ *
  * * `storage_path` - storage path for persistence engine.
  *
  * # Safety
@@ -297,8 +298,7 @@ enum norddrop_result norddrop_reject_file(const struct norddrop *dev,
  */
 enum norddrop_result norddrop_start(const struct norddrop *dev,
                                     const char *listen_addr,
-                                    const char *config,
-                                    const char *tracker_context);
+                                    const char *config);
 
 /**
  * Stop norddrop instance